+++ scripts/generate_openapi_spec.py
@@ -58,7 +58,7 @@ def build_openapi_spec(publish):
         title=package,
         version=ver,
         info=dict(description=desc),
-        plugins=("apispec.ext.flask",),
+        plugins=("apispec.ext.flask", "apispec.ext.marshmallow"),
     )
 
     # Add marshmallow schemas to the specification here